finish

+++ app/src/main/java/uz/turgunboyevjurabek/rizon/madels/UserMain/ProductSalesData2.kt
@@ -4,5 +4,5 @@ data class ProductSalesData2(
     val name: String,
     val product_sale_amount: Int,
     val product_sale_summa: Int,
-    val product_sales_percent: Int
+    val product_sales_percent: Double
 )
